# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insta API",
        "description": "API документация для проекта Insta",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "2. Profile"
                ],
                "summary": "Get current user profile",
                "description": "Returns information about the authenticated user",
                "operationId": "49d94eefc8e488baccc0aca063463849",
                "responses": {
                    "200": {
                        "description": "Current user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "john_doe"
                                        },
                                        "subscribers": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "publications": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://yourdomain.com/storage/avatars/avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "2. Profile"
                ],
                "summary": "Update current user profile",
                "description": "Updates the authenticated user's profile information",
                "operationId": "cf55cbb67aa3ac5ef6f6cf0cec3c0ccb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "john_doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe Updated"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "newjohn@example.com"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "john_doe_updated"
                                        },
                                        "subscribers": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "publications": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://yourdomain.com/storage/avatars/avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field must be a valid email address."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field must be a valid email address."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/avatar": {
            "post": {
                "tags": [
                    "2. Profile"
                ],
                "summary": "Upload or update user avatar",
                "description": "Uploads a new avatar for the authenticated user",
                "operationId": "49393366d7a1c4dec8723b4369ebd7a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "Avatar image file (jpeg or png, max 1MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "john_doe"
                                        },
                                        "subscribers": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "publications": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://yourdomain.com/storage/avatars/new_avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The avatar must be a file of type: jpeg, png."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "avatar": [
                                                    "The avatar must be a file of type: jpeg, png."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Paste your Sanctum access token here",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}